restructure and migrate search index
