+++ src/modules/search_index/actions.js
@@ -1,4 +1,4 @@
-import * as YoutubeVideoAPI from 'util/youtube_video_util';
+import * as YoutubeVideoAPI from './util';
 
 export const RECEIVE_VIDEOS = 'RECEIVE_VIDEOS';
 export const CLEAR_VIDEOS = 'CLEAR_VIDEOS';
@@ -13,7 +13,6 @@ export const clearVideos = () => ({
 });
 
 export const searchVideos = (query, nextPageToken, pageNumber = 1) => dispatch => {
-
   return YoutubeVideoAPI.fetchVideos(query, nextPageToken).then(
     res => res.json()
   ).then(
@@ -25,9 +24,9 @@ export const searchVideos = (query, nextPageToken, pageNumber = 1) => dispatch =
            for (let i = 0; i < videos.items.length; i++) {
              videos.items[i]['statistics'] = videoStatResults.items[i].statistics;
            }
- 
+
            videos['query'] = query;
- 
+
            return dispatch(receiveVideos(videos));
          }
        );
